+++ src/Router.js
@@ -2,8 +2,8 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import GlobalStyles from './styles/GlobalStyles';
 import { ThemeProvider } from 'styled-components';
 import theme from './styles/theme';
-import Category from './pages/Category/Category';
-import Detail from './pages/Detail/Detail';
+import Category from './pages/category/Category';
+import Detail from './pages/detail/Detail';
 
 function Router() {
   return (
