+++ src/pages/detail/Detail.js
